First refactor of projectors into separate class.
