+++ mbirjax/__init__.py
@@ -1,5 +1,6 @@
+from .projectors import *
 from .tomography_model import *
+from .parallel_beam import *
 from .vcd_utils import *
 from .gpu_memory import *
-from .parallel_beam import *
